// apps/proyecto-integrado/src/app/state/actions/trips/trips-action.types.enum.ts
export enum TripsActionTypes {
  CREATE_HOTEL_REQUEST = '[Trips] Create Hotel Request',
  CREATE_HOTEL_SUCCESS = '[Trips] Create Hotel Success',
  CREATE_HOTEL_FAILURE = '[Trips] Create Hotel Failure',
  GET_HOTELS_REQUEST = '[Trips] Get Hotels Request',
  GET_HOTELS_SUCCESS = '[Trips] Get Hotels Success',
  GET_HOTELS_FAILURE = '[Trips] Get Hotels Failure',
  DELETE_HOTEL_REQUEST = '[Trips] Delete Hotel Request',
  DELETE_HOTEL_SUCCESS = '[Trips] Delete Hotel Success',
  DELETE_HOTEL_FAILURE = '[Trips] Delete Hotel Failure',
  UPDATE_HOTEL_REQUEST = '[Trips] Update Hotel Request',
  UPDATE_HOTEL_SUCCESS = '[Trips] Update Hotel Success',
  UPDATE_HOTEL_FAILURE = '[Trips] Update Hotel Failure',

  CREATE_DESTINATION_REQUEST = '[Trips] Create Destination Request',
  CREATE_DESTINATION_SUCCESS = '[Trips] Create Destination Success',
  CREATE_DESTINATION_FAILURE = '[Trips] Create Destination Failure',
  GET_DESTINATIONS_REQUEST = '[Trips] Get Destinations Request',
  GET_DESTINATIONS_SUCCESS = '[Trips] Get Destinations Success',
  GET_DESTINATIONS_FAILURE = '[Trips] Get Destinations Failure',
  DELETE_DESTINATION_REQUEST = '[Trips] Delete Destination Request',
  DELETE_DESTINATION_SUCCESS = '[Trips] Delete Destination Success',
  DELETE_DESTINATION_FAILURE = '[Trips] Delete Destination Failure',
  UPDATE_DESTINATION_REQUEST = '[Trips] Update Destination Request',
  UPDATE_DESTINATION_SUCCESS = '[Trips] Update Destination Success',
  UPDATE_DESTINATION_FAILURE = '[Trips] Update Destination Failure',

  CREATE_TRANSPORT_REQUEST = '[Trips] Create Transport Request',
  CREATE_TRANSPORT_SUCCESS = '[Trips] Create Transport Success',
  CREATE_TRANSPORT_FAILURE = '[Trips] Create Transport Failure',
  GET_TRANSPORTS_REQUEST = '[Trips] Get Transports Request',
  GET_TRANSPORTS_SUCCESS = '[Trips] Get Transports Success',
  GET_TRANSPORTS_FAILURE = '[Trips] Get Transports Failure',
  DELETE_TRANSPORT_REQUEST = '[Trips] Delete Transport Request',
  DELETE_TRANSPORT_SUCCESS = '[Trips] Delete Transport Success',
  DELETE_TRANSPORT_FAILURE = '[Trips] Delete Transport Failure',
  UPDATE_TRANSPORT_REQUEST = '[Trips] Update Transport Request',
  UPDATE_TRANSPORT_SUCCESS = '[Trips] Update Transport Success',
  UPDATE_TRANSPORT_FAILURE = '[Trips] Update Transport Failure',

  CREATE_TRIP_REQUEST = '[Trips] Create Trip Request',
  CREATE_TRIP_SUCCESS = '[Trips] Create Trip Success',
  CREATE_TRIP_FAILURE = '[Trips] Create Trip Failure',
  GET_TRIPS_REQUEST = '[Trips] Get Trips Request',
  GET_TRIPS_SUCCESS = '[Trips] Get Trips Success',
  GET_TRIPS_FAILURE = '[Trips] Get Trips Failure',
  DELETE_TRIP_REQUEST = '[Trips] Delete Trip Request',
  DELETE_TRIP_SUCCESS = '[Trips] Delete Trip Success',
  DELETE_TRIP_FAILURE = '[Trips] Delete Trip Failure',
  UPDATE_TRIP_REQUEST = '[Trips] Update Trip Request',
  UPDATE_TRIP_SUCCESS = '[Trips] Update Trip Success',
  UPDATE_TRIP_FAILURE = '[Trips] Update Trip Failure',
}